requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
pydantic[email]>=2.7.2
pydantic-settings>=2.5.2
supabase==2.4.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
Pillow==10.3.0
python-magic==0.4.27
qrcode[pil]==7.4.2
httpx==0.27.0
anthropic==0.26.0
slowapi==0.1.9
pytest==8.2.0
pytest-asyncio==0.23.6
